feat: volar labs support
