+++ packages/language-server/bin/nodeServer.js
@@ -3,5 +3,5 @@ if (process.argv.includes('--version')) {
 	const pkgJSON = require('../package.json');
 	console.log(`${pkgJSON['version']}`);
 } else {
-	require('../dist/index.js');
+	require('../dist/nodeServer.js');
 }
