+++ packages/language-server/package.json
@@ -4,7 +4,7 @@
   "author": "withastro",
   "license": "MIT",
   "type": "commonjs",
-  "main": "out/index.js",
+  "main": "dist/index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/withastro/language-tools",
@@ -26,32 +26,32 @@
     "test:match": "pnpm run test -g"
   },
   "dependencies": {
-    "@astrojs/compiler": "^1.3.2",
+    "@astrojs/compiler": "^1.4.1",
     "@jridgewell/sourcemap-codec": "^1.4.15",
-    "@volar/language-core": "^1.5.0",
-    "@volar/language-server": "^1.5.0",
-    "@volar/language-service": "^1.5.0",
-    "@volar/source-map": "^1.5.0",
-    "muggle-string": "^0.2.2",
-    "volar-service-css": "^0.0.0",
-    "volar-service-emmet": "^0.0.0",
-    "volar-service-html": "^0.0.0",
-    "volar-service-prettier": "^0.0.0",
-    "volar-service-typescript": "^0.0.0",
-    "volar-service-typescript-twoslash-queries": "^0.0.0",
-    "vscode-html-languageservice": "^5.0.4",
+    "@volar/language-core": "^1.6.1",
+    "@volar/language-server": "^1.6.1",
+    "@volar/language-service": "^1.6.1",
+    "@volar/source-map": "^1.6.1",
+    "muggle-string": "^0.3.1",
+    "volar-service-css": "^0.0.2",
+    "volar-service-emmet": "^0.0.2",
+    "volar-service-html": "^0.0.2",
+    "volar-service-prettier": "^0.0.1",
+    "volar-service-typescript": "^0.0.2",
+    "volar-service-typescript-twoslash-queries": "^0.0.2",
+    "vscode-html-languageservice": "^5.0.5",
     "vscode-uri": "^3.0.7",
     "fast-glob": "^3.2.12"
   },
   "devDependencies": {
-    "@astrojs/svelte": "^2.1.0",
+    "@astrojs/svelte": "^2.1.1",
     "@astrojs/vue": "^2.1.1",
-    "@types/chai": "^4.3.0",
-    "@types/mocha": "^9.1.1",
+    "@types/chai": "^4.3.5",
+    "@types/mocha": "^10.0.1",
     "@types/node": "^16.18.26",
-    "astro": "^2.3.0",
-    "chai": "^4.3.6",
-    "mocha": "^9.2.2",
+    "astro": "^2.4.4",
+    "chai": "^4.3.7",
+    "mocha": "^10.2.0",
     "tsx": "^3.12.7",
     "typescript": "~5.0.4",
     "vscode-languageserver-protocol": "^3.17.3",
