+++ packages/language-server/src/core/svelte.ts
@@ -29,6 +29,7 @@ class SvelteFile implements VirtualFile {
 	fileName: string;
 	mappings!: Mapping<FileRangeCapabilities>[];
 	embeddedFiles!: VirtualFile[];
+	codegenStacks = [];
 
 	constructor(public sourceFileName: string, public snapshot: ts.IScriptSnapshot) {
 		this.fileName = sourceFileName;
