+++ packages/language-server/src/nodeServer.ts
@@ -1,0 +1,4 @@
+import { createConnection, startLanguageServer } from '@volar/language-server/node';
+import { plugin } from './languageServerPlugin.js';
+
+startLanguageServer(createConnection(), plugin);
