+++ packages/vscode/package.json
@@ -361,12 +361,14 @@
     "test:grammar": "pnpm build:grammar && node ./test/grammar/test.mjs",
     "update-grammar-snapshots": "node ./test/grammar/test.mjs --updateSnapshot"
   },
+  "dependencies": {
+    "@astrojs/language-server": "workspace:*"
+  },
   "devDependencies": {
-    "@astrojs/language-server": "workspace:*",
     "@types/node": "^16.18.26",
     "@types/vscode": "1.67.0",
-    "@volar/language-server": "1.4.1",
-    "@volar/vscode": "1.4.1",
+    "@volar/language-server": "1.6.1",
+    "@volar/vscode": "1.6.1",
     "@vscode/vsce": "latest",
     "esbuild": "^0.17.18",
     "esbuild-plugin-copy": "^2.1.1",
